% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/census-boundaries.R
\name{census_boundaries}
\alias{cb_2014_us_cd114_20m}
\alias{cb_2014_us_county_20m}
\alias{cb_2014_us_state_20m}
\alias{census_boundaries}
\title{Cartographic boundary files from the U.S. Census Bureau}
\description{
The U.S. Census Bureau provides
\href{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}{cartographic
boundary files} for current U.S. boundaries. The datasets in this package
provide a selection of boundaries from the 2014 Census files. They are
objects of class \code{\link[sp]{SpatialPolygonsDataFrame}}. The attributes
data associated with these boundaries is unchanged from what is available in
the Census boundaries, with the exception that a \code{state_name} column has
been added where necessary for convience in filtering the boundaries. See the
U.S. Census Bureau's
"\href{https://www.census.gov/geo/reference/geoidentifiers.html}{Understanding
Geographic Identifiers (GEOIDs)}" and their
"\href{https://www.census.gov/geo/reference/geocodes.html}{Geographic Codes}"
pages for the details of this attribute data.
}
\references{
U.S. Census Bureau,
  \href{https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html}{Cartographic
   Boundary Shapefiles} (2014).
}

