% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-data-pkg.R
\name{install_USAboundariesData}
\alias{install_USAboundariesData}
\title{Install the \code{USAboundariesData} package after checking with the user}
\usage{
install_USAboundariesData()
}
\description{
Install the \code{USAboundariesData} package after checking with the user
}
