% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_plane.R
\name{state_plane}
\alias{state_plane}
\title{State Plane Projections}
\usage{
state_plane(state, plane_id = NULL, type = c("epsg", "proj4"))
}
\arguments{
\item{state}{The state code.}

\item{plane_id}{The state plane geographic zone identifier. A
\code{NULL} value will return a projection for the entire state.}

\item{type}{A string identifying the output as either an EPSG code or PROJ4 string.}
}
\value{
A string
}
\description{
Get the State Plane projection for a specific state or region within a state.
}
\examples{
state_plane (state = "MA", type = "epsg")
state_plane (state = "MA", plane_id = "mainland", type = "proj4")

}
\seealso{
For documentation of the underlying state plane projection data frame, see \code{\link{state_proj}}.
}
