% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{states_contemporary_lores}
\alias{states_contemporary_lores}
\title{U.S. state boundaries}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 52 rows and 13 columns.
}
\usage{
states_contemporary_lores
}
\description{
The U.S. Census Bureau provides
\href{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}{cartographic boundary files}
for current U.S. boundaries. This package has only the
low-resolution contemporary state boundaries. Other census boundary files are
provided by and documented in the USAboundariesData package.
}
\seealso{
For citations for the other Census boundary files provided by the
  USAboundariesData package, see the \code{census_boundaries} documentation
  in that package.
}
\keyword{datasets}
