% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/usboundaries-package.r
\docType{package}
\name{USAboundaries}
\alias{USAboundaries}
\alias{USAboundaries-package}
\title{USAboundaries: Historical boundaries of the United States of America,
1629-2000}
\description{
This package provides spatial objects with the boundaries of states or
counties in the United States of America from 1629 to 2000. It provides data
from the \href{http://publications.newberry.org/ahcbp/}{Atlas of Historical
County Boundaries}.
}
\details{
The copyright to the data used in this package is owned by the Newberry
Library, and it is included in this package under the terms of the
\href{http://creativecommons.org/licenses/by-nc-sa/2.5/}{Creative Commons
Attribution-NonCommercial-ShareAlike 2.5 Generic} (CC BY-NC-SA 2.5) license.
}
\seealso{
us_boundaries
}

