% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{usboundaries}
\alias{usboundaries}
\alias{usboundaries-package}
\title{usboundaries: Historical boundaries of the United States, 1629--2000}
\description{
This package provides spatial objects with the boundaries of states or
counties in the United States of America from 1629 to 2000. It provides data
from the \href{http://publications.newberry.org/ahcbp/}{Atlas of Historical
County Boundaries}.
}
\details{
The copyright in the data used in this package is owned by the Newberry
Library, and it is included in this package under the terms of the Creative
Commons Attribution-NonCommercial-ShareAlike 2.5 Generic (CC BY-NC-SA 2.5)
license.
}

