% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-data-pkg.R
\name{check_data_package}
\alias{check_data_package}
\title{Check whether to install USAboundariesData and install if necessary}
\usage{
check_data_package()
}
\value{
No return value: if USAboundariesData is not installed or a newer
  version is required, a console message will appear asking to install
  USAboundariesData. Else nothing will return.
}
\description{
If the \code{USAboundariesData} package is not installed, install it from the
}
