% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-data-pkg.R
\name{install_USAboundariesData}
\alias{install_USAboundariesData}
\title{Install the \code{USAboundariesData} package after checking with the user}
\usage{
install_USAboundariesData()
}
\value{
No return value: if USAboundariesData is not installed, it will ask
  the user if they want to install it. If user selects "yes" then it will be
  installed. If no, abort.
}
\description{
Install the \code{USAboundariesData} package after checking with the user
}
