% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/us_boundaries.R
\name{us_boundaries}
\alias{us_boundaries}
\title{Get state/county boundaries of the United States on a specified date.}
\usage{
us_boundaries(map_date = NULL, type = c("state", "county", "congressional"),
  resolution = c("20m"), states = NULL)
}
\arguments{
\item{map_date}{The date of the map as some object coercable to a date with
\code{as.Date()}. For state maps the date must be between 3 September 1783
and 31 December 2000; for county maps the date must be between 30 December
1636 and 31 December 2000.}

\item{type}{The type of the map, either \code{"state"} or \code{"county"}.}

\item{resolution}{The resolution of the map.}

\item{states}{A character vector of state or territory names. Only boundaries
inside these states/territories will be return.}
}
\value{
A \code{SpatialPolygonsDataFrame} object.
}
\description{
This function takes a date and returns the county or state boundaries of the
United States for that date.
}
\examples{
map_states   <- us_boundaries("1850-07-04")
map_counties <- us_boundaries(as.Date("1850-07-04"), type = "county")
new_england <- c("Massachusetts", "Vermont", "New Hampshire", "Connecticut",
                 "Rhode Island", "Maine")
map_ne <- us_boundaries("1943-02-12", states = new_england)

if (require(sp)) {
  plot(map_states)
  plot(map_counties)
  plot(map_ne)
}
}

