% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_zipcodes.R
\name{us_zipcodes}
\alias{us_zipcodes}
\title{Zip Code Tabulation Areas (contemporary)}
\usage{
us_zipcodes()
}
\value{
An \code{sf} object.
}
\description{
Get the current (2019) centroids for U.S Zipcode Tabulation Areas from the
U.S. Census Bureau. The centroids were calculated from the ZCTA boundary
files available on the U.S. Census Bureau website.
}
\examples{
if (require(USAboundariesData, quietly = TRUE)) {
  us_zipcodes()
}

}
\seealso{
For documentation of and citation to the underlying shapefiles for
  contemporary data from the U.S. Census Bureau, see the
  \code{census_boundaries} documentation in the USAboundariesData package.
}
